added xml for solo picture, initial buttons design
added a 'solo_picture.xml' file to display a big, single picture. This is to be implemented in the future when a small-version of an image (called a 'thumbnail') from the picture fragment is pressed on, that thumbnail will transition to this frame, showing the whole big image, the title, as well as the option to go back, to set as background or 'next'. The placement (the design) of buttons inside 'solo_picture.xml' is only draft, will be adjusted if needed in the future