Add 'set-background' functionality, adjust button to CompatButton, improve buttons' UI
Changed the UI a bit, making it look nicer by adding colors to the button; still needs more beautification; used AppCompatButton instead of Button for better compatibility with old version (especially this, we're asked to develop with the minSdk = 29 - API = 29 - Android 10.0)
Added the functionality to set the home screen as the current image, when exiting the application, the home screen still persists. The option to set image as the lock screen may be another button (another functionality rather than included inside the 'set background' functionality).
After this, need to add animation for transition to the next and previous images ('Next' and 'Previous' buttons).
About the 'set background' functionality, maybe need to add a loading screen when the image is being processed? And/or allow user to choose angle, position of the image to be set?